Add default Guidebox API key
Guidebox has expanded usability of default permissions. It used to be 1 access per second, which would cripple a multi-user API key with navigation like this provides.  However, they've since increased to several hundred per second.  The monthly limit is still 10,000 fetches, so that may prove to be a problem in the future.